refactor: shard out uploading frontend logic in EmailRecipients component
