+++ frontend/src/components/common/index.ts
@@ -20,3 +20,4 @@ export { default as ConfirmModal } from './confirm-modal'
 export { default as CredLabelInput } from './cred-label-input'
 export { default as Dropdown } from './dropdown'
 export { default as BodyWrapper } from './body-wrapper'
+export { default as CsvUpload } from './csv-upload'
