+++ frontend/src/services/email.service.ts
@@ -1,24 +1,5 @@
 import axios, { AxiosError } from 'axios'
-
-interface PresignedUrlResponse {
-  presignedUrl: string
-  transactionId: string
-}
-
-interface CsvStatusResponse {
-  isCsvProcessing: boolean
-  csvFilename?: string
-  tempCsvFilename?: string
-  csvError?: string
-  numRecipients?: number
-  preview?: PreviewMessage
-}
-
-interface PreviewMessage {
-  body: string
-  subject: string
-  replyTo: string | null
-}
+import { EmailPreview } from 'classes'
 
 export async function saveTemplate(
   campaignId: number,
@@ -69,87 +50,9 @@ export async function sendPreviewMessage({
   }
 }
 
-export async function getPresignedUrl({
-  campaignId,
-  mimeType,
-}: {
-  campaignId: number
-  mimeType: string
-}): Promise<PresignedUrlResponse> {
-  try {
-    const response = await axios.get(
-      `/campaign/${campaignId}/email/upload/start`,
-      {
-        params: {
-          mime_type: mimeType,
-        },
-      }
-    )
-    const {
-      transaction_id: transactionId,
-      presigned_url: presignedUrl,
-    } = response.data
-    return { transactionId, presignedUrl } as PresignedUrlResponse
-  } catch (e) {
-    errorHandler(e, 'Error completing file upload')
-  }
-}
-
-export async function completeFileUpload({
-  campaignId,
-  transactionId,
-  filename,
-}: {
-  campaignId: number
-  transactionId: string
-  filename: string
-}): Promise<void> {
-  try {
-    await axios.post(`/campaign/${campaignId}/email/upload/complete`, {
-      transaction_id: transactionId,
-      filename,
-    })
-    return
-  } catch (e) {
-    errorHandler(e, 'Error completing file upload')
-  }
-}
-
-export async function getCsvStatus(
-  campaignId: number
-): Promise<CsvStatusResponse> {
-  try {
-    const response = await axios.get(
-      `/campaign/${campaignId}/email/upload/status`
-    )
-    const {
-      is_csv_processing: isCsvProcessing,
-      csv_filename: csvFilename,
-      temp_csv_filename: tempCsvFilename,
-      csv_error: csvError,
-      num_recipients: numRecipients,
-      preview,
-    } = response.data
-    const result = {
-      isCsvProcessing,
-      csvFilename,
-      tempCsvFilename,
-      csvError,
-      numRecipients,
-    } as CsvStatusResponse
-    if (preview) {
-      const { subject, body, reply_to: replyTo } = preview
-      result.preview = { subject, body, replyTo }
-    }
-    return result
-  } catch (e) {
-    errorHandler(e, 'Error getting csv status')
-  }
-}
-
 export async function getPreviewMessage(
   campaignId: number
-): Promise<PreviewMessage> {
+): Promise<EmailPreview> {
   try {
     const response = await axios.get(`/campaign/${campaignId}/email/preview`)
     const { body, subject, reply_to: replyTo } = response.data?.preview
